+++ application/models/Tracker/Tracker_Tag_Model.php
@@ -21,7 +21,7 @@ class Tracker_Tag_Model extends Tracker_Base_Model {
 		return $success;
 	}
 
-	public function getMalID(int $userID, int $titleID) : ?int {
+	public function getMalID(int $userID, int $titleID) {
 		$query = $this->db->select('tags')
 		                    ->where('user_id', $userID)
 		                    ->where('title_id', $titleID)
@@ -29,9 +29,9 @@ class Tracker_Tag_Model extends Tracker_Base_Model {
 
 		$malID = NULL;
 		if($query->num_rows() > 0 && ($tag_string = $query->row()->tags) && !is_null($tag_string)) {
-			$arr   = preg_grep('/^mal:([0-9]+)$/', explode(',', $tag_string));
+			$arr   = preg_grep('/^mal:([0-9]+|none)$/', explode(',', $tag_string));
 			if(!empty($arr)) {
-				$malID = (int) explode(':', $arr[0])[1];
+				$malID = explode(':', $arr[0])[1];
 			}
 		}
 		return $malID;